+++ spec/misc-spec.js
@@ -37,32 +37,6 @@ describe("fix()", function() {
 
 });
 
-describe("extract()", function() {
-
-  var collection = {
-    tic: {blah:"foo",yada:"qwerty"},
-    tac: {blah:"foo",hoot:"asdf"},
-    toe: {blah:"bar",burp:"zxcv"},
-    tug: {hoot:"12345"}
-  }
-
-  var hash = new Firehash();
-  hash.extract("blah",collection);
-
-  it("pulls out an object with keys 'foo','bar'", function() {
-    expect(hash.keys()).toEqual(['foo','bar']);
-  });
-
-  it("has a 'foo' value", function() {
-    expect(Object.keys(hash.get("foo"))).toEqual(["tic","tac"]);
-  });
-
-  it("has a 'bar' value", function() {
-    expect(Object.keys(hash.get("bar"))).toEqual(["toe"]);
-  });
-
-});
-
 describe("forEach()", function() {
 
   var collection = {