+++ docs/changelog.md
@@ -1,4 +1,8 @@
-**2016-10-20** _v1.1_
+**2016-10-21** _v1.1.1_
+
+* `extract()`; Apply `transformFn`.
+
+**2016-10-21** _v1.1_
 
 * Remove `import()` and `export()`. Incorporate their functionality into `set()` and `get()`.
 * Build out tests.